Refactor WCS Parameters parsing to fit WMS and WMTS
