+++ ogc-example/src/main/scala/geotrellis/server/ogc/wcs/WcsView.scala
@@ -16,7 +16,8 @@
 
 package geotrellis.server.ogc.wcs
 
-import geotrellis.server.ogc.wcs.params._
+import geotrellis.server.ogc.params.ParamError
+
 import com.typesafe.scalalogging.LazyLogging
 import org.backuity.ansi.AnsiFormatter.FormattedHelper
 import org.http4s.scalaxml._
@@ -24,16 +25,15 @@ import org.http4s._
 import org.http4s.dsl.io._
 import cats.effect._
 import cats.data.Validated
-import java.net._
 
-import org.http4s.headers.`Content-Type`
+import java.net._
 
 class WcsView(wcsModel: WcsModel, serviceUrl: URL) extends LazyLogging {
 
   private def handleError[Result](result: Either[Throwable, Result])(implicit ee: EntityEncoder[IO, Result]) = result match {
     case Right(res) =>
       logger.info("response", res.toString)
-      Ok(res, `Content-Type`(new MediaType("image", "tiff", false, true, List("tif", "tiff"))))
+      Ok(res)
     case Left(err) =>
       logger.error(s"error: $err", err)
       InternalServerError(err.toString)
@@ -44,7 +44,7 @@ class WcsView(wcsModel: WcsModel, serviceUrl: URL) extends LazyLogging {
   def responseFor(req: Request[IO])(implicit cs: ContextShift[IO]): IO[Response[IO]] = {
     WcsParams(req.multiParams) match {
       case Validated.Invalid(errors) =>
-        val msg = WcsParamsError.generateErrorMessage(errors.toList)
+        val msg = ParamError.generateErrorMessage(errors.toList)
         logger.warn(s"""Error parsing parameters: ${msg}""")
         BadRequest(s"""Error parsing parameters: ${msg}""")
